refactor: use atoms provider
